backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
numpy==1.26.4
onnxruntime==1.16.3
Pillow==10.1.0
colorlog==6.8.0
prometheus-client==0.19.0
mlflow==2.9.2
